requirements.txt
appnope==0.1.4
asttokens==2.4.1
blinker==1.8.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
dominate==2.9.1
exceptiongroup==1.2.1
executing==2.0.1
firebase-admin==6.5.0
Flask==2.2.2
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.0.2
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.5.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.53.0
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.20.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-secret-manager==2.16.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-oauth2-tool==0.0.3
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.0
grpcio-status==1.62.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
msgpack==1.0.8
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pyzmq==26.0.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.18
visitor==0.1.3
wcwidth==0.2.13
Werkzeug==2.2.2
WTForms==3.1.2
